新しいフックポイントに対応
