+++ View/entry_confirm_add_mailmaga.twig
@@ -19,6 +19,7 @@ You should have received a copy of the GNU General Public License
 along with this program; if not, write to the Free Software
 Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #}
+<div class="dl_table">
 <dl>
 <dt>{{ form_label(form.mailmaga_flg) }}</dt>
 <dd>
@@ -27,3 +28,4 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
     </div>
 </dd>
 </dl>
+</div>
