+++ event.yml
@@ -1,6 +1,19 @@
-#
-# Event fock
-#
+# New hookpoint for version >= 3.0.9
+Entry/index.twig:
+    - [onRenderEntryIndex, NORMAL]
+Entry/confirm.twig:
+    - [onRenderEntryConfirm, NORMAL]
+front.entry.index.complete:
+    - [onFrontEntryIndexComplete, NORMAL]
+Mypage/change.twig:
+    - [onRenderMypageChange, NORMAL]
+front.mypage.change.index.complete:
+    - [onFrontMypageChangeIndexComplete, NORMAL]
+Admin/Customer/edit.twig:
+    - [onRenderAdminCustomerEdit, NORMAL]
+admin.customer.edit.index.complete:
+    - [onAdminCustomerEditIndexComplete, NORMAL]
+# deprecated event (for v3.0.0 - 3.0.8)
 eccube.event.render.entry.before:
     - [onRenderEntryBefore, NORMAL]
 eccube.event.controller.entry.after:
